refactor(rust): simplify mailbox architecture
Mailbox used to be a concrete type because we needed to be able to
re-queue messages after calling `ctx.receive()`.  However as part of
the transport architecture we added `ctx.forward` which can now be
used by the `Cancel` type to re-queue messages.

This allows us to remove the Mailbox type and integrate it more
closely with the Context.  This removes one internal routing hop when
sending messages to a worker, thus decreasing complexity and messaging
delay.